backend/pyproject.toml
[tool.poetry]
name = "golf-league-backend"
version = "0.1.0"
description = "Golf League Management Backend API"
authors = ["Ben Christian <bchristian83@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.116.2,<0.117.0"
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
sqlalchemy = ">=2.0.43,<3.0.0"
alembic = ">=1.16.5,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-jose = {extras = ["cryptography"], version = ">=3.5.0,<4.0.0"}
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
python-multipart = ">=0.0.20,<0.0.21"
pydantic-settings = ">=2.10.1,<3.0.0"
pydantic = {extras = ["email"], version = "^2.11.9"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.18.1"
requests = "^2.32.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
